+++ elevenlabs-convai/CHANGELOG.md
@@ -6,10 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - `termination_reason` to `conversation::Metadata`
+- `send_tool_result` to `AgentWebSocket`
 ### Changed
 - **Breaking**: `ElevenLabsAgentClient` name to `AgentWebSocket`
-- **Breaking**: `start_conversation` method to `start_session` (perhaps not the methods)
-- **Breaking**: `stop_conversation` method to `end_session`
+
 ## [0.2.0] - 2025-02-28
 ### Added
 - dependency on elevenlabs_rs 0.5.1